// src/shared/api/generated/model/campaignCompensationType.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Marz API
 * Backend de Marz. Fuente de verdad del contrato HTTP entre marz-api y marz-front.
Se consume con `oapi-codegen` (server) y `openapi-typescript` + `openapi-fetch` (client).

 * OpenAPI spec version: 0.1.0
 */

export type CampaignCompensationType = typeof CampaignCompensationType[keyof typeof CampaignCompensationType];


export const CampaignCompensationType = {
  payment: 'payment',
  product_trade: 'product_trade',
  payment_plus_product: 'payment_plus_product',
} as const;